Moved GameStateChecker methods to GameFieldData. Added access modifiers for DataSerialization class
