+++ GameOfLife/GameData/GameFieldData.cs
@@ -2,12 +2,17 @@
 {
     /// <summary>
     /// Contains data of array dimensions and data about 
-    /// next generation results based on previous
+    /// next generation results based on previous. Also contains data of  live 
+    /// cell count in current iterations and iteration numbering.
     /// </summary>
+    /// 
+
     [Serializable]
     public class GameFieldData
     {
         public int[,]? gameFieldArray;
+        public static int iterationCount = 1;
+
         public GameFieldData()
         {
             int rows = UserComunicator.GetInputValueRanged(Repository.EnterRowsCountMessage, 5, 50);
@@ -105,6 +110,40 @@
                     gameField[i, j] = nextGenerationOfCells[i, j];
                 }
             }
+        }
+
+        /// <summary>
+        /// Calculates alive cells in GameField
+        /// </summary>
+        /// <param name="gameFieldArray">Current game array</param>
+        /// <returns>Alive cell count</returns>
+        public static int AliveCellsInCurrentIteration(int[,] gameFieldArray)
+        {
+            int aliveCells = 0;
+
+            for (int i = 0; i < gameFieldArray.GetLength(0); i++)
+            {
+                for (int j = 0; j < gameFieldArray.GetLength(1); j++)
+                {
+                    if (gameFieldArray[i, j] == 1)
+                    {
+                        aliveCells++;
+                    }
+                }
+            }
+
+            return aliveCells;
+        }
+
+
+        /// <summary>
+        /// Iteration counter
+        /// </summary>
+        /// <returns>incremented iteration count by 1</returns>
+        public static int IterationCounter()
+        {
+            iterationCount++;
+            return iterationCount;
         }
     }
 }
