+++ GameOfLife/GameData/SavedObjects.cs
@@ -5,10 +5,10 @@ namespace GameOfLife
     /// Objects to serialize in binary format.
     /// </summary>
     [Serializable]
-    public class SavedObjects 
+    public class SavedObjects
     {
         public GameFieldData? GameFieldData;
-        public int Iteration = GameStateChecker.iterationCount;
+        public int Iteration = GameFieldData.iterationCount;
     }
-   
+
 }
